==> 01-Vectores.Rmd <==
# Vectores, matrices, listas y data frames 

```{r, include=FALSE}
knitr::opts_chunk$set(
  comment = "#>", fig.width = 6,
  eval=FALSE, message=FALSE, warning=FALSE, include=TRUE
)
```

## P1 Vectores {-}

(a) Generar un vector que contenga la siguiente secuencia de números: $1,2,3,4,...,10$
  
```{r}
seq(1,10,1) #forma 1
1:10 #forma 2
seq(10) #forma 3
c(1,2,3,4,5,6,7,8,9,10) #forma 4
```
  

(b) Asigne el vector creado a $x$, determine su tipo y dimensión


```{r}
x <- seq(1, 10, 1)
typeof(x)
length(x)
```


(c) Calcule la suma y el producto de los elementos del vector $x$

```{r}
sum(x)
prod(x)
```

(d) Calcule la suma acumulada y el producto acumulado de los elementos del vector $x$

```{r}
cumsum(x)
cumprod(x)
```


(e) Construya un vector lógico a partir de los elementos del vector x que son divisibles por 2 ¿Cuántos números son divisibles por 2? ¿Cuáles son divisibles por 2?

```{r}
y <- x %% 2 == 0
sum(x %% 2 == 0)
x[x %% 2 == 0]
```


(f) Genere el siguiente vector: “mes_1”, “mes_2”,. . . , “mes_10”

```{r}
paste0("mes_", x)
```

## P2 Vectores {-}

Genere un vector que repita roja cuatro veces y azul tres veces.

```{r}
c(rep("roja", 4), rep("azul", 3)) # Forma 1
rep(c("roja", "azul"), c(4, 3)) # Forma 2

```


## P3 Tablas {-}

```{r eval = TRUE, include = TRUE, echo=FALSE}
library(tibble)
library(kableExtra)
tibble(Nombre = c("Pedro", "Ana", "Maya", "Max", "Paula"), 
       Edad = c(17, NA, 23, NA, 20)) %>% 
  kbl() %>% 
  kable_classic(full_width = F)
```


(a) A partir de esta tabla genere un vector $x$ con las edades de los estudiantes. Los nombres de cada uno de los elementos del vector deben ser los nombres de los estudiantes.

```{r}
x = c("Pedro"=17, "Ana"=NA, "Maya"=23, "Max"=NA, "Paula"=20)
```

(b) Defina un vector $y$ que excluya los missing values (NA).

```{r}
y = x[-c(2,4)] #forma 1
y=x[c(-2,-4)] # forma 2
y=x[c("Pedro","Maya", "Paula")] # forma 3
y=x[!is.na(x)] # forma 4
y = na.omit(x) # forma 5
```

(c) Asigne las edades de Pedro y Ana a un vector $z$.

```{r}
z <- x[c("Pedro","Ana")] #forma 1
z <- x[c(1,2)] #forma 2
```


(d) Calcule la media aritmética de las edades de los 5 estudiantes usando la función mean y la fórmula para el cálculo de la media aritmética:

$$
\overline{x} = \frac{\sum^n_{i=1}x_i}{n}
$$

```{r}
mean(x) # ¿Qué obtiene?
mean(x, na.rm = TRUE) #forma 1
sum(x[!is.na(x)])/length(x[!is.na(x)]) #forma 2
```

(e) Calcule la desviación estándar de las edades de los 5 estudiantes usando la función sd y la fórmula para el cálculo de la desviación estándar:

$$
s = \sqrt{\frac{\sum^n_{i=1}(x_i-\overline{x})^2}{n-1}}
$$

```{r}
sd(x, na.rm = TRUE) # Forma 1
sqrt(sum((x[!is.na(x)]-mean(x, na.rm = TRUE))^2)/(length(x[!is.na(x)]) - 1)) # Forma 2
sqrt(sum((na.omit(x)-mean(x, na.rm = TRUE))^2)/(length(na.omit(x)) - 1)) # Forma 3
```


(g) Calcule la proporción de estudiantes que tienen 20 años o menos. 

```{r}
mean(x<=20, na.rm = TRUE)
```

(h) Calcule el promedio de edad de los estudiantes cuyas edades son 20 años o menos. 

```{r}
mean(x[x<=20], na.rm = TRUE) #Forma 1
mean(x[which(x<=20)]) # Forma 2
```

(i) Escriba summary(x) en el script ¿Qué obtiene?

```{r}
summary(x)
```

## P4 Tasa crecimiento {-}

Calcule la tasa de crecimiento promedio anual del PIB si las tasas son: 20% y −20%

```{r}
 r_g <- c(20,-20)
(prod(1 + r_g/100)^(1/length(r_g)) - 1)*100
```

## P5 Tasa crecimiento {-}

Si una variable crece a un $x%$ por período, construya una fórmula que le permita obtener el número de períodos para que la variable duplique su valor. Calcule el número de períodos si las tasas de crecimiento son 4%, 6%, 8% y 10%.

```{r}
x = c(4, 6, 8, 10)
log(2)/log(1+(x/100))
```

## P6 Valor presente {-}

(a) Calcule el valor presente de un flujo de efectivo de $1000 por un total de diez períodos. Asuma una tasa de descuento de 10%.

El valor presente $(VP)$ es igual a:

$$
VP = \sum^{T}_{t=1}\frac{F_t}{(1+r)^t} = \sum^{T}_{t=1}F_t\times (1+r)^{-t}
$$


```{r}
r = 0.1 #tasa de descuento
flujo_efectivo = 1000 #flujo por período
n=10
periodo = 1:n #períodos
```

```{r}
valor_presente <- sum(flujo_efectivo * (1 + r) ^ -periodo)
paste0("Valor Presente", "=","$", round(valor_presente,0)) #forma 1
paste0("Valor Presente", "=","$", round((flujo_efectivo/r)*(1 - (1 + r)^-n),0)) #forma 2
```

(b) Calcule el valor presente neto de un flujo de efectivo de \$300, \$1200, \$1000 y una inversión inicial de $900. Asuma que las tasas de descuento para cada uno de los períodos son 10%, 20%, 15%, respectivamente.

$$
VPN = -900+\frac{100}{1.1}+\frac{1200}{1.1\times 1.2}+ \frac{1000}{1.1\times 1.2\times 1.15} = 940.58
$$

```{r}
# Tasa de interés
r <- c(0.1, 0.2, 0.15)

# Factores de descuento
yearly_discount_factors <- (1 + r) ^ - 1
yearly_discount_factors
discount_factors <- c(1 , cumprod(yearly_discount_factors))
discount_factors

# Definir vector de flujos
cash_flow <- c(-900, 300, 1200, 1000)

# Calcular valor presente
VPN <- sum(cash_flow * discount_factors)
paste0("Valor Presente Neto", "=","$", round(VPN,0))
```

(c) Calcule el número de períodos en que se recupera la inversión (payback period) del flujo de efectivo de \$300, \$1200, \$1000 y una inversión inicial de \$900.

```{r}
paste0("Payback period = ", min(which(cumsum(cash_flow) >= 0)) - 1)
```

## P7 Muestras {-}


\begin{align*}
skewness &= \frac{\frac{\sum^n_{i=1}(x_i-\overline{x})^3}{n}}{\left[\frac{\sum^n_{i=1}(x_i-\overline{x})^2}{n}  \right]^{\frac{3}{2}}}\\
kurtosis &= \frac{\frac{\sum^n_{i=1}(x_i-\overline{x})^4}{n}}{\left[\frac{\sum^n_{i=1}(x_i-\overline{x})^2}{n}  \right]^{2}}\\
\end{align*}


Genere una muestra de 10000 números aleatorios provenientes de la distribución normal estándar y calcule la skewness y kurtosis, usando las siguientes fórmulas:


```{r}
z = rnorm(10000)

skewness <- (sum((z - mean(z))^3)/length(z))/(sum((z - mean(z))^2)/length(z))^1.5

kurtosis <- (sum((z - mean(z))^4)/length(z))/(sum((z - mean(z))^2)/length(z))^2

moments::skewness(z)
moments::kurtosis(z)
```

## P8 Matrices {-}

Sea X una matriz :
$$
X = \begin{bmatrix}
0 & 10 & -5\\
5 & 8 & -4\\
0 & 1 & -1
\end{bmatrix}
$$

(a) Construya la matriz X y asigne nombres a sus filas y columnas.
 
```{r}
a <- c(0,10,-5)
b <- c(5,8,-4)
c <- c(0,1,-1)


X <- rbind(a,b,c) # Forma 1
rownames(X) <- c("fila1", "fila2", "fila3")
colnames(X) <- c("columna1", "columna2", "columna3")

cbind(c(0,5,0), c(10,8,1),c(-5,-4,-1)) # Forma 2
e = c(0,10,-5, 5,8,-4, 0,1,-1)
matrix(e, nrow=3, ncol=3, byrow = TRUE) # Forma 3

```
 
(b) Almacene la segunda fila de X en el vector $z$.

```{r}
z <- X[2,]
```

(c) Almacene los elementos (2, 3) y (3, 3) en el vector $w$.
```{r}
w <- X[c(2,3),3]
```


## P9 Listas {-}

(a) Genere una lista que consista del vector a igual a \{8, 12, 16,. . . , 36\}, un vector b que contenga las palabras hola y ¿Cómo estás? y un elemento c que contiene el número 5.

```{r}
lista_A = list(a=seq(8, 36, 4), b=list("hola","¿Cómo estás?"), c = 5)
```

(b) Obtenga el primer elemento de la lista.

```{r}
lista_A[["a"]] # Forma 1
lista_A$a # Forma 2
lista_A["a"] # Forma 3
```

(c) Obtenga los primeros dos elementos de la lista.

```{r}
lista_A[1:2]
```

(d) Asigne "¿Cómo estás?" a s

```{r}
lista_A[["b"]][2] # Forma 1
lista_A[["b"]][[2]] # Forma 2
```

## P10 Manipulación {-}

La siguiente tabla presenta la distribución por zona de una muestra, separando entre hombres y mujeres:

```{r eval = TRUE, include = TRUE, echo=FALSE, message=FALSE, warning=FALSE}
zona <- c("A", "B", "C", "D")
hombres <- c(800,200,700,300)
mujeres <- c(700,300,600,400)
muestra_total_df <- data.frame(zona, hombres, mujeres) # dataframe
muestra_total_df %>% kbl %>% kable_classic(full_width = F)
```


(a) Construya dicha tabla como un data frame.

```{r}
zona <- c("A", "B", "C", "D")
hombres <- c(800,200,700,300)
mujeres <- c(700,300,600,400)
muestra_total_df <- data.frame(zona, hombres, mujeres) # dataframe
```


(b) Construya dicha tabla como un tibble.

```{r}
muestra_total_tb <- tibble::tibble(zona = c("A", "B", "C", "D"),
                                   hombres = c(800,200,700,300),
                                   mujeres = c(700,300,600,400))

```

(c) Escriba en el script $summary(muestra_total_tb)$ ¿Qué obtiene?

```{r}
summary(muestra_total_tb)
```

(d) Escriba en el script $str(muestra_total_tb)$ y $dplyr::glimpse(muestra_total_tb)$ ¿Qué obtiene?

```{r}
str(muestra_total_tb)
dplyr::glimpse(muestra_total_tb)
```

(e) Presente las primeras tres y últimas dos observaciones del $muestra_total_tb$.

```{r}
library(dplyr)
muestra_total_tb %>% slice_head(n = 3)
muestra_total_tb %>% slice_tail(n = 2)
```

(f) Extraer la columna $mujeres$.

```{r}
muestra_total_tb$mujeres # Forma 1
muestra_total_tb %>% pull(mujeres) # Forma 2
```

(g) Construya una tabla con las observaciones de las variables $zona$ y $mujeres$.

```{r}
muestra_total_tb %>% select(-hombres) # forma 1
muestra_total_tb %>% select(zona, mujeres) # forma 2
```

(h) Construya una tabla con las muestras de las zonas $A$ y $C$.

```{r}
muestra_total_tb %>% filter(zona=="A" | zona=="C") # Forma 1
muestra_total_tb %>% filter(zona %in% c("A","C")) # Forma 2
```

(i) Construya una variable $muestra_total$ que sea igual a la suma de $hombres$ y $mujeres$ y una variable $prop_muestra$ que representa la proporción de la muestra total de una determinada región.

```{r}
muestra_total_tb <- muestra_total_tb %>%
  mutate(mayor = if_else(muestra_total > 1000, "alto", "bajo"))
muestra_total_tb
```

(j) Genere una variable que toma el valor “alto” si la muestra total de una zona determinada es mayor a 1000 e igual a “bajo”, si no son mayores.

```{r}
 muestra_total_tb <- muestra_total_tb %>%
  mutate(mayor = if_else(muestra_total > 1000, "alto", "bajo"))
muestra_total_tb
```

(k) Guarde muestra_total_tb con extensión $.csv$.

```{r}
readr::write_csv(muestra_total_tb, file = "datos/muestra_total.csv")
```
